allow create chatroom from map page and pass router param to chatpage
