+++ src/router/index.js
@@ -112,7 +112,8 @@ const beforeRouteEnter = async (to, from, next) => {
 
   if (to.meta.userRole) {
     // Redirect user if user role does not match
-    if (to.meta.userRole !== Vue.$store.getters['auth/getUserRole']) { // TODO: FIX async BUG
+    if (to.meta.userRole !== Vue.$store.getters['auth/getUserRole']) {
+      // TODO: FIX async BUG
       next({ name: 'home' });
     } else next();
   } else next();
